+++ alembic.ini
@@ -5,7 +5,7 @@ script_location = migrations
 # the 'revision' command, regardless of autogenerate
 # revision_environment = false
 
-sqlalchemy.url = postgresql://robert@/lexibank
+sqlalchemy.url = postgresql://postgres@/lexibank
 
 [production]
 script_location = migrations
