+++ setup.cfg
@@ -6,6 +6,12 @@ addopts =
     --cov=lexibank
     --cov-report term-missing
 
+[coverage:run]
+source = tsammalex
+
+[coverage:report]
+show_missing = true
+
 [compile_catalog]
 directory = lexibank/locale
 domain = clld
