+++ setup.py
@@ -20,15 +20,20 @@ setup(
     include_package_data=True,
     zip_safe=False,
     install_requires=[
+        'clld~=4.0',
         'clldmpg~=3.1',
         'clld-glottologfamily-plugin>=2.0.1',
         'pycldf',
+        'sqlalchemy',
+        'waitress',
     ],
     extras_require={
-        'dev': ['flake8', 'waitress'],
-        'test': [
+        'dev': [
+            'flake8',
             'psycopg2',
-            'tox',
+            'tox'
+        ],
+        'test': [
             'mock',
             'pytest>=3.1',
             'pytest-clld',
