+++ tox.ini
@@ -1,0 +1,11 @@
+[tox]
+envlist =
+    py{27,34,35,36}
+skip_missing_interpreters =
+    true
+
+[testenv]
+extras =
+    test
+commands =
+    python -m pytest -m "not selenium" {posargs}
